+++ food-for-thought/app/home.tsx
@@ -3,7 +3,6 @@ import { router } from "expo-router";
 import { Card, Text, Icon } from '@rneui/themed';
 import SearchBar from "@/components/SearchBar";
 import { TouchableOpacity } from "react-native-gesture-handler";
-import Header from "@/components/Header";
 import RecommendedRestaurant from "@/components/RecommendedRestaurant";
 import MapView, { Marker } from "react-native-maps";
 import { styles } from '../styles/app-styles'; 
@@ -12,13 +11,14 @@ import axios from "axios";
 import Constants from 'expo-constants';
 import { Restaurant } from "@/app/map";
 import Layout from "@/components/Layout";
+import SafetyWarning from "@/components/SafetyWarning";
 
 // Component
 const Home = () => {
   const [fetchedRestaurants, setFetchedRestaurants] = useState<Restaurant[]>([]);
+  const [showModal, setShowModal] = useState<boolean>(true);
 
   const HOST_IP = Constants.expoConfig?.extra?.HOST_IP;
-
   const pic = require('../assets/images/react-logo.png'); // placeholder restaurant image
 
   // Fetch specific restaurants by ID
@@ -45,6 +45,10 @@ const Home = () => {
     fetchRestaurants();
   }, []);
 
+  const handleNotedPress = () => {
+    setShowModal(false);
+  };
+
   const handleSearch = (search: string) => {
     // Temp fix for onSearch error. Potentially route to map.tsx and parse searchFilter.
     return '';
@@ -63,8 +67,9 @@ const Home = () => {
 
   return (
     <Layout>
+      {showModal && <SafetyWarning handleNotedPress={ handleNotedPress } />} 
       {/* Card for the Restaurant finder */}
-      <TouchableOpacity onPress={() => router.push('/warning')}>
+      <TouchableOpacity onPress={() => router.push('/map')}>
         <Card containerStyle={styles.finderCard}>
           <View>
             <SearchBar onSearch={handleSearch} />
