+++ server/.env_template
@@ -4,3 +4,4 @@ BACKEND_URL=
 FOURSQUARE_API_KEY=
 JWT_SECRET=
 HOST_IP=
+GPT_KEY=
